admin/requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
datatables
gunicorn
